scripts/root_requirements.txt
anthropic>=0.80.0
openai>=2.21.0
python-dotenv>=1.2.1
requests>=2.32.5
beautifulsoup4>=4.14.0
httpx>=0.28.0
